feat(web): create simulation registry/loader with caching
- Create SimulationLoader class to handle all simulation operations
- Add manifest loading with caching for performance optimization
- Add individual simulation loading with caching
- Refactor YAML parsing logic into dedicated loader class
- Replace inline simulation loading with centralized registry
- Add proper TypeScript interfaces for type safety
- Clean separation of concerns from main application logic
